Add dismissable banner to legacy frontend with link to try the new site
The banner is behind a 'try-new-frontend' waffle, which defaults to off
for the moment.

A few yaks had to be shaved to achieve all this:
- The banner code was separated and called from zamboni/global.js, in
  order to be shown on non-impala pages (i.e. themes).
- The banner code was modified to try to only display one banner at a
  time.
- Because the banners were not shown in non-impala pages before, some
  tweaks to the layout css properties were made, to make sure the banner
  have the correct margin everywhere.
- The mobile site link handling code prevents default to avoid following
  a link to '#'.
