+++ src/olympia/browse/tests.py
@@ -12,6 +12,7 @@ from django.utils.translation import trim_whitespace
 import pytest
 import mock
 from pyquery import PyQuery as pq
+from waffle.testutils import override_switch
 
 from olympia import amo
 from olympia.amo.tests import TestCase
@@ -78,6 +79,14 @@ class TestListing(TestCase):
         super(TestListing, self).setUp()
         cache.clear()
         self.url = reverse('browse.extensions')
+
+    def test_try_new_frontend_banner_presence(self):
+        response = self.client.get(self.url)
+        assert 'AMO is getting a new look.' not in response.content
+
+        with override_switch('try-new-frontend', active=True):
+            response = self.client.get(self.url)
+            assert 'AMO is getting a new look.' in response.content
 
     def test_default_sort(self):
         r = self.client.get(self.url)
@@ -1052,6 +1061,14 @@ class TestPersonas(TestCase):
             a._current_version = v
             a.status = amo.STATUS_PUBLIC
             a.save()
+
+    def test_try_new_frontend_banner_presence(self):
+        response = self.client.get(self.landing_url)
+        assert 'AMO is getting a new look.' not in response.content
+
+        with override_switch('try-new-frontend', active=True):
+            response = self.client.get(self.landing_url)
+            assert 'AMO is getting a new look.' in response.content
 
     def test_does_not_500_in_development(self):
         with self.settings(DEBUG=True):
